numpyfy knotvector
