+++ tests/edges/bsplinecurve_objects/bounding_box_contour.json
