+++ volmdlr/nurbs/fitting.py
@@ -19,6 +19,7 @@ code for any particular purpose.
 import cython.cimports.libc.math as math_c
 import cython
 import numpy as np
+from numpy.typing import NDArray
 from scipy.linalg import lu_factor, lu_solve
 
 from cython.cimports.libcpp.vector import vector
@@ -52,7 +53,7 @@ def interpolate_curve(points: np.ndarray[np.double_t, ndim == 2], degree: cython
     u_k: cython.double[:] = compute_params_curve(points, centripetal)
 
     # Compute knot vector
-    knotvector: vector[cython.double] = compute_knot_vector(degree, num_points, u_k)
+    knotvector: cython.double[:] = compute_knot_vector(degree, num_points, u_k)
 
     # Do global interpolation
     matrix_a: cython.double[:, :] = _build_coeff_matrix(degree, knotvector, u_k, num_points)
@@ -61,7 +62,7 @@ def interpolate_curve(points: np.ndarray[np.double_t, ndim == 2], degree: cython
     lu_matrix, piv = lu_factor(matrix_a)
     ctrlpts: np.ndarray[np.double_t, ndim == 2] = lu_solve((lu_matrix, piv), points)
 
-    knots = np.unique(knotvector)
+    knots: np.ndarray[np.double_t, ndim == 1] = np.unique(knotvector)
     knot_multiplicities = [core.find_multiplicity(knot, knotvector) for knot in knots]
 
     return ctrlpts, knots, knot_multiplicities
@@ -100,8 +101,8 @@ def interpolate_surface(points, size_u: cython.int, size_v: cython.int, degree_u
     u_k, v_l = compute_params_surface(points, size_u, size_v, use_centripetal)
 
     # Compute knot vectors
-    kv_u: vector[cython.double] = compute_knot_vector(degree_u, size_u, u_k)
-    kv_v: vector[cython.double] = compute_knot_vector(degree_v, size_v, v_l)
+    kv_u: cython.double[:] = compute_knot_vector(degree_u, size_u, u_k)
+    kv_v: cython.double[:] = compute_knot_vector(degree_v, size_v, v_l)
 
     j: cython.int
     u: cython.int
@@ -168,7 +169,7 @@ def approximate_curve(points, degree: cython.int, **kwargs):
     u_k: np.ndarray[np.double_t, ndim == 1] = compute_params_curve(points, use_centripetal)
 
     # Compute knot vector
-    knotvector: vector[cython.double] = compute_knot_vector2(degree, num_dpts, num_cpts, u_k)
+    knotvector: cython.double[:] = compute_knot_vector2(degree, num_dpts, num_cpts, u_k)
 
     # Compute matrix N
     matrix_n: np.ndarray[np.double_t, ndim == 2] = np.zeros((num_dpts - 2, num_cpts - 2), dtype=np.float64)
@@ -263,8 +264,8 @@ def approximate_surface(points, size_u: cython.int, size_v: cython.int, degree_u
     u_k, v_l = compute_params_surface(points, size_u, size_v, use_centripetal)
 
     # Compute knot vectors
-    kv_u: vector[cython.double] = compute_knot_vector2(degree_u, size_u, num_cpts_u, u_k)
-    kv_v: vector[cython.double] = compute_knot_vector2(degree_v, size_v, num_cpts_v, v_l)
+    kv_u: cython.double[:] = compute_knot_vector2(degree_u, size_u, num_cpts_u, u_k)
+    kv_v: cython.double[:] = compute_knot_vector2(degree_v, size_v, num_cpts_v, v_l)
 
     matrix_nu: np.ndarray[np.double_t, ndim == 2] = np.zeros((size_u - 2, num_cpts_u - 2), dtype=np.float64)
     i: cython.int
@@ -370,7 +371,7 @@ def approximate_surface(points, size_u: cython.int, size_v: cython.int, degree_u
 @cython.boundscheck(False)
 def compute_knot_vector(
     degree: cython.int, num_points: cython.size_t, params: cython.double[:]
-) -> vector[cython.double]:
+) -> cython.double[:]:
     """
     Computes a knot vector from the parameter list using averaging method.
 
@@ -385,24 +386,28 @@ def compute_knot_vector(
     :return: knot vector
     :rtype: list
     """
-    # Start knot vector
-    knotvector: vector[cython.double] = vector[cython.double](degree + 1, 0.0)
-    temp_kv: cython.double
+    knotvector: vector[cython.double] = vector[cython.double](num_points + degree + 1, 0.0)
     i: cython.size_t
     j: cython.size_t
+    degree_float: cython.double = cython.cast(cython.double, degree)
     # Use averaging method (Eq 9.8) to compute internal knots in the knot vector
+
     for i in range(num_points - degree - 1):
-        temp_kv = (1.0 / float(degree)) * sum(params[j] for j in range(i + 1, i + degree + 1))
-        knotvector.push_back(temp_kv)
+        knotvector[degree + 1 + i] = (1.0 / degree_float) * sum(params[j] for j in range(i + 1, i + degree + 1))
 
     # End knot vector
-    for _ in range(degree + 1):
-        knotvector.push_back(1.0)
+    for i in range(num_points, num_points + degree + 1):
+        knotvector[i] = 1.0
 
-    return knotvector
+    return np.asarray(knotvector, dtype=np.float64)
 
 
-def compute_knot_vector2(degree, num_dpts, num_cpts, params):
+@cython.cfunc
+@cython.cdivision
+@cython.wraparound(False)
+@cython.boundscheck(False)
+def compute_knot_vector2(degree: cython.int, num_dpts: cython.int,
+                         num_cpts: cython.int, params: cython.double[:]) -> cython.double[:]:
     """
     Computes a knot vector ensuring that every knot span has at least one :math:`\\overline{u}_{k}`.
 
@@ -420,19 +425,22 @@ def compute_knot_vector2(degree, num_dpts, num_cpts, params):
     :rtype: list
     """
     # Start knot vector
-    knotvector = [0.0 for _ in range(degree + 1)]
+    knotvector: np.ndarray[np.double_t, ndim == 1] = np.zeros(num_cpts + degree + 1, dtype=np.float64)
 
     # Compute "d" value - Eq 9.68
-    d = float(num_dpts) / float(num_cpts - degree)
+    d: cython.double = cython.cast(cython.double, num_dpts) / cython.cast(cython.double, num_cpts - degree)
+    i: cython.int
+    j: cython.int
+    alpha: cython.double
+    temp_kv: cython.double
     # Find internal knots
     for j in range(1, num_cpts - degree):
-        i = int(j * d)
+        i = cython.cast(cython.int, (j * d))
         alpha = (j * d) - i
         temp_kv = ((1.0 - alpha) * params[i - 1]) + (alpha * params[i])
-        knotvector.append(temp_kv)
+        knotvector[degree + j] = temp_kv
 
-    # End knot vector
-    knotvector += [1.0 for _ in range(degree + 1)]
+    knotvector[num_cpts:] = 1.0
 
     return knotvector
 
@@ -469,7 +477,7 @@ def compute_params_curve(points: np.ndarray[np.double_t, ndim == 2], centripetal
     # Divide individual chord lengths by the total chord length
     u_k: np.ndarray[np.double_t, ndim == 1] = np.zeros(num_points, dtype=np.double)
     for i in range(num_points):
-        u_k[i] = np.sum(cds[0 : i + 1]) / d
+        u_k[i] = np.sum(cds[0: i + 1]) / d
 
     return u_k
 
@@ -538,7 +546,7 @@ def compute_params_surface(points: np.ndarray[np.double_t, ndim == 2], size_u: c
 
 @cython.cfunc
 def _build_coeff_matrix(
-    degree: cython.int, knotvector: vector[cython.double], params: cython.double[:], num_points: cython.size_t
+    degree: cython.int, knotvector: cython.double[:], params: cython.double[:], num_points: cython.size_t
 ) -> cython.double[:, :]:
     """
     Builds the coefficient matrix for global interpolation.
