+++ volmdlr/nurbs/operations.py
@@ -578,7 +578,7 @@ def decompose_curve(obj, return_params: bool = False, **kwargs):
     params = []
     umin, umax = obj.domain
     param_start = umin
-    while knots:
+    while knots.any():
         knot = knots[0]
         curves = split_curve(curve, param=knot, **kwargs)
         multi_curve.append(curves[0])
@@ -675,7 +675,7 @@ def helper_decompose(srf, idx, split_func, return_params, **kws):
             param_min, param_max = domain[2], domain[3]
     param_start = param_min
     params = []
-    while knots:
+    while knots.any():
         knot = knots[0]
         srfs = split_func(srf, param=knot, **kws)
         srf_list.append(srfs[0])
